backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "langchain>=0.3.27",
    "langchain-community>=0.3.27",
    "langchain-openai>=0.3.28",
    "openai>=1.97.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "uvicorn>=0.35.0",
]
